feat(privhelper): bit-packed address scheme (12-bit tag, /28, 16-bit v6 ctx)
Repartition the addressing — a /24 per kennel was wasteful (nothing wants
254 loopback addresses), so reclaim those bits for more users.

- IPv4: `127 | tag(12) | ctx(8) | host(4)` → a /28 per kennel (16 addrs).
  A user owns the /20 their 12-bit tag selects: 4096 users, 256 v4-enabled
  kennels each. `tag` grows u8 -> u16; V4_PREFIX 24 -> 28.
- IPv6: `fd | gid(40) | ctx(16) | host(64)` → /64 per kennel. The random
  per-user gid isolates users, so no tag in v6. `ctx` is 16-bit; its low 8
  bits coincide with the v4 ctx (dual-stack kennels share one context;
  v6-only kennels can use the full 16-bit range). `ctx` grows u8 -> u16
  through AddrRequest and the wire Request (now 294 bytes: a reserved byte
  + u16 ctx).
- alloc: tag field parses 0..=4095 (rejects out-of-range); gid unchanged.

validate tests rewritten with v4()/v6() constructors for the packed
layout, plus new cases: a v4-capped kennel (ctx 300) has no v4 address; a
v6 kennel with ctx 300 validates. Root integration test adds/removes the
computed /28 alias (127.0.144.81 for tag 9/ctx 5). 34 unit + integration
+ root tests green. Authoritative scheme doc (07-3-network.md) updated.

Follow-up: the policy address-substitution syntax (`127.<tag>.<ctx>.1`
template literals) needs reworking — addresses are now computed, not
octet-aligned; that lands with the policy resolver.

+++ crates/kennel-privhelper/src/client.rs
@@ -67,7 +67,7 @@ pub fn delete_cgroup(helper: &Path, path: PathBuf) -> io::Result<Response> {
 /// # Errors
 ///
 /// As [`invoke`].
-pub fn add_address(helper: &Path, ctx: u8, interface: &str, addr: IpAddr, prefix: u8) -> io::Result<Response> {
+pub fn add_address(helper: &Path, ctx: u16, interface: &str, addr: IpAddr, prefix: u8) -> io::Result<Response> {
     invoke(helper, &addr_request(Op::AddAddr, ctx, interface, addr, prefix))
 }
 
@@ -76,7 +76,7 @@ pub fn add_address(helper: &Path, ctx: u8, interface: &str, addr: IpAddr, prefix
 /// # Errors
 ///
 /// As [`invoke`].
-pub fn del_address(helper: &Path, ctx: u8, interface: &str, addr: IpAddr, prefix: u8) -> io::Result<Response> {
+pub fn del_address(helper: &Path, ctx: u16, interface: &str, addr: IpAddr, prefix: u8) -> io::Result<Response> {
     invoke(helper, &addr_request(Op::DelAddr, ctx, interface, addr, prefix))
 }
 
@@ -91,7 +91,7 @@ const fn cgroup_request(op: Op, path: PathBuf) -> Request {
     }
 }
 
-fn addr_request(op: Op, ctx: u8, interface: &str, addr: IpAddr, prefix: u8) -> Request {
+fn addr_request(op: Op, ctx: u16, interface: &str, addr: IpAddr, prefix: u8) -> Request {
     Request {
         op,
         ctx,
